backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
sarvamai
google-genai
requests
